// backend/src/common/inversify-containers/types.ts
export enum IocConainerTypes {
  Env = "Env",
  MongoDB = "MongoDB",
  Logging = "Logging",
  UserService = "UserService",
  UserDomain = "UserDomain",
  DepartmentDomain = "DepartmentDomain",
  ActionLogService = "ActionLogService",
  DepartmentService = "DepartmentService",
  UserRepository = "UserRepository",
  DepartmentRepository = "DepartmentRepository",
  ActionLogRepository = "ActionLogRepository",
  UserController = "UserController",
  DepartmentController = "DepartmentController",
  ActionLogController = "ActionLogController",
  AuthController = "AuthController",
  PassportService = "PassportService",
  ActionLogDomain = "ActionLogDomain"
}